feat(duckdb): make push local-only, keep Quack literalization for reads
