+++ internal/duckdb/push.go
@@ -2,12 +2,10 @@ package duckdb
 
 import (
 	"context"
-	"crypto/rand"
 	"crypto/sha256"
 	"database/sql"
 	"encoding/hex"
 	"fmt"
-	"reflect"
 	"slices"
 	"sort"
 	"strconv"
@@ -133,8 +131,8 @@ func duckPricingUpsertStatement(prices []db.ModelPricing) (string, []any) {
 }
 
 func (s *Sync) listDuckModelPricing(ctx context.Context) ([]db.ModelPricing, error) {
-	rows, err := queryDuckDBContext(
-		ctx, s.duck, s.connectionKind, s.quack,
+	rows, err := s.duck.QueryContext(
+		ctx,
 		`SELECT model_pattern, input_per_mtok,
 			output_per_mtok, cache_creation_per_mtok,
 			cache_read_per_mtok, updated_at
@@ -593,669 +591,12 @@ func projectInSyncScope(project string, projects, excludeProjects []string) bool
 func (s *Sync) execMutation(
 	ctx context.Context, exec duckMutationExecutor, stmt string, args ...any,
 ) error {
-	if s.connectionKind != duckDBQuackClientConnection {
-		_, err := exec.ExecContext(ctx, stmt, args...)
-		return err
-	}
-	if _, ok := exec.(*duckRemoteMutationBatch); ok {
-		_, err := exec.ExecContext(ctx, stmt, args...)
-		return err
-	}
-	// Quack attachments can accept plain inserts, but DELETE, UPDATE, and
-	// ON CONFLICT are planned against proxy storage and currently fail with
-	// GetStorageInfo errors. Run those mutations on the server-side base DB.
-	sqlText, err := duckSQLWithArgs(stmt, args...)
-	if err != nil {
-		return err
-	}
-	_, err = exec.ExecContext(ctx, "FROM "+quackAttachmentName+".query(?)", sqlText)
+	_, err := exec.ExecContext(ctx, stmt, args...)
 	return err
 }
 
 type duckMutationExecutor interface {
 	ExecContext(context.Context, string, ...any) (sql.Result, error)
-}
-
-const (
-	duckRemoteMutationCoalesceMaxBytes = 2 << 20
-	duckRemoteInsertCoalesceMaxRows    = 256
-	duckRemoteInsertCoalesceMaxBytes   = 1 << 20
-)
-
-type duckRemoteMutationBatch struct {
-	statements              []string
-	renderedValid           bool
-	renderedStatementsCache []duckRemoteRenderedStatement
-	renderedStatementBytes  int
-}
-
-type duckRemoteRenderedStatement struct {
-	sql        string
-	start, end int
-}
-
-func (b *duckRemoteMutationBatch) ExecContext(
-	_ context.Context, stmt string, args ...any,
-) (sql.Result, error) {
-	sqlText, err := duckSQLWithArgs(stmt, args...)
-	if err != nil {
-		return nil, err
-	}
-	b.statements = append(b.statements, sqlText)
-	b.invalidateRendered()
-	return duckNoopResult{}, nil
-}
-
-func (b *duckRemoteMutationBatch) Len() int {
-	return len(b.statements)
-}
-
-func (b *duckRemoteMutationBatch) appendBatch(other *duckRemoteMutationBatch) {
-	if other == nil || other.Len() == 0 {
-		return
-	}
-	oldLen := len(b.statements)
-	preserveRendered := b.renderedValid
-	rendered := b.renderedStatementsCache
-	renderedBytes := b.renderedStatementBytes
-	b.statements = append(b.statements, other.statements...)
-	if !preserveRendered {
-		b.invalidateRendered()
-		return
-	}
-	for _, stmt := range other.rendered() {
-		stmt.start += oldLen
-		stmt.end += oldLen
-		rendered = append(rendered, stmt)
-		renderedBytes += len(stmt.sql)
-	}
-	b.renderedStatementsCache = rendered
-	b.renderedStatementBytes = renderedBytes
-	b.renderedValid = true
-}
-
-func (b *duckRemoteMutationBatch) invalidateRendered() {
-	b.renderedValid = false
-	b.renderedStatementsCache = nil
-	b.renderedStatementBytes = 0
-}
-
-type duckNoopResult struct{}
-
-func (duckNoopResult) LastInsertId() (int64, error) { return 0, nil }
-func (duckNoopResult) RowsAffected() (int64, error) { return 0, nil }
-
-func (s *Sync) execRemoteMutationBatch(
-	ctx context.Context, label string, batch *duckRemoteMutationBatch,
-) error {
-	return execDuckRemoteMutationBatch(
-		ctx, s.execRemoteSQLRetry, label, batch, true,
-	)
-}
-
-func appendDuckRemoteMutationBatch(
-	ctx context.Context,
-	execCoalesced func(context.Context, string) error,
-	label string,
-	current *duckRemoteMutationBatch,
-	next *duckRemoteMutationBatch,
-	maxBytes int,
-) (*duckRemoteMutationBatch, error) {
-	if current == nil {
-		current = &duckRemoteMutationBatch{}
-	}
-	if next == nil || next.Len() == 0 {
-		return current, nil
-	}
-	if maxBytes <= 0 {
-		maxBytes = duckRemoteMutationCoalesceMaxBytes
-	}
-	nextBytes := next.transactionBytes()
-	if nextBytes > maxBytes {
-		if err := execDuckRemoteMutationBatch(
-			ctx, execCoalesced, label, current, true,
-		); err != nil {
-			return current, err
-		}
-		return &duckRemoteMutationBatch{}, fmt.Errorf(
-			"duckdb remote mutation batch exceeds remote mutation coalesce budget: %d > %d",
-			nextBytes, maxBytes,
-		)
-	}
-	if current.Len() > 0 && current.combinedTransactionBytes(next) > maxBytes {
-		if err := execDuckRemoteMutationBatch(ctx, execCoalesced, label, current, true); err != nil {
-			return current, err
-		}
-		current = &duckRemoteMutationBatch{}
-	}
-	current.appendBatch(next)
-	return current, nil
-}
-
-func execDuckRemoteMutationBatch(
-	ctx context.Context,
-	exec func(context.Context, string) error,
-	label string,
-	batch *duckRemoteMutationBatch,
-	coalesce bool,
-) (err error) {
-	if batch.Len() == 0 {
-		return nil
-	}
-	if coalesce {
-		if err := exec(ctx, batch.transactionSQL()); err != nil {
-			if isDuckRemoteMutationTimeoutError(err) {
-				return err
-			}
-			if rollbackErr := exec(ctx, "ROLLBACK"); rollbackErr != nil {
-				return fmt.Errorf("%w; rollback %s: %v", err, label, rollbackErr)
-			}
-			return err
-		}
-		return nil
-	}
-	return execDuckRemoteMutationStatements(ctx, exec, label, batch.statements)
-}
-
-func execDuckRemoteMutationRenderedStatements(
-	ctx context.Context,
-	exec func(context.Context, string) error,
-	label string,
-	batch *duckRemoteMutationBatch,
-) error {
-	if batch.Len() == 0 {
-		return nil
-	}
-	rendered := batch.rendered()
-	statements := make([]string, 0, len(rendered))
-	for _, stmt := range rendered {
-		statements = append(statements, stmt.sql)
-	}
-	return execDuckRemoteMutationStatements(ctx, exec, label, statements)
-}
-
-func execDuckRemoteMutationStatements(
-	ctx context.Context,
-	exec func(context.Context, string) error,
-	label string,
-	statements []string,
-) (err error) {
-	if len(statements) == 0 {
-		return nil
-	}
-	if err := exec(ctx, "BEGIN TRANSACTION"); err != nil {
-		return fmt.Errorf("begin %s: %w", label, err)
-	}
-	needsRollback := true
-	defer func() {
-		if !needsRollback {
-			return
-		}
-		rollbackErr := exec(ctx, "ROLLBACK")
-		if err != nil && rollbackErr != nil {
-			err = fmt.Errorf("%w; rollback %s: %v", err, label, rollbackErr)
-		}
-	}()
-	for i, stmt := range statements {
-		if err := exec(ctx, stmt); err != nil {
-			return fmt.Errorf(
-				"execute %s statement %d/%d: %w",
-				label, i+1, len(statements), err,
-			)
-		}
-	}
-	if err := exec(ctx, "COMMIT"); err != nil {
-		return fmt.Errorf("commit %s: %w", label, err)
-	}
-	needsRollback = false
-	return nil
-}
-
-func execDuckRemoteMutationBatchWithStatementFallback(
-	ctx context.Context,
-	execCoalesced func(context.Context, string) error,
-	execStatement func(context.Context, string) error,
-	label string,
-	batch *duckRemoteMutationBatch,
-) error {
-	if err := execDuckRemoteMutationBatch(ctx, execCoalesced, label, batch, true); err != nil {
-		if ctx.Err() != nil ||
-			isStaleQuackConnectionError(err) ||
-			isDuckRemoteMutationTimeoutError(err) {
-			return err
-		}
-		return execDuckRemoteMutationBatch(ctx, execStatement, label, batch, false)
-	}
-	return nil
-}
-
-func execDuckRemoteMutationBatchOversizeWithStatementFallback(
-	ctx context.Context,
-	execCoalesced func(context.Context, string) error,
-	execStatement func(context.Context, string) error,
-	label string,
-	batch *duckRemoteMutationBatch,
-	maxBytes int,
-) error {
-	if maxBytes <= 0 {
-		maxBytes = duckRemoteMutationCoalesceMaxBytes
-	}
-	if batch.transactionBytes() > maxBytes {
-		return execDuckRemoteMutationRenderedStatements(
-			ctx, execStatement, label, batch,
-		)
-	}
-	return execDuckRemoteMutationBatchWithStatementFallback(
-		ctx, execCoalesced, execStatement, label, batch,
-	)
-}
-
-func (b *duckRemoteMutationBatch) transactionSQL() string {
-	var sqlText strings.Builder
-	statements := b.rendered()
-	sqlText.WriteString("BEGIN TRANSACTION;\n")
-	for _, stmt := range statements {
-		sqlText.WriteString(stmt.sql)
-		sqlText.WriteString(";\n")
-	}
-	sqlText.WriteString("COMMIT")
-	return sqlText.String()
-}
-
-func (b *duckRemoteMutationBatch) transactionBytes() int {
-	if b == nil || b.Len() == 0 {
-		return 0
-	}
-	statements := b.rendered()
-	return duckRemoteTransactionBytesFor(
-		b.renderedStatementBytes, len(statements),
-	)
-}
-
-func duckRemoteTransactionBytesFor(statementBytes, statementCount int) int {
-	if statementCount == 0 {
-		return 0
-	}
-	return len("BEGIN TRANSACTION;\n") +
-		statementBytes +
-		len(";\n")*statementCount +
-		len("COMMIT")
-}
-
-func (b *duckRemoteMutationBatch) combinedTransactionBytes(
-	other *duckRemoteMutationBatch,
-) int {
-	if b == nil || b.Len() == 0 {
-		return other.transactionBytes()
-	}
-	if other == nil || other.Len() == 0 {
-		return b.transactionBytes()
-	}
-	left := b.rendered()
-	right := other.rendered()
-	return duckRemoteTransactionBytesFor(
-		b.renderedStatementBytes+other.renderedStatementBytes,
-		len(left)+len(right),
-	)
-}
-
-func (b *duckRemoteMutationBatch) rendered() []duckRemoteRenderedStatement {
-	if b == nil || b.Len() == 0 {
-		return nil
-	}
-	if b.renderedValid {
-		return b.renderedStatementsCache
-	}
-	b.renderedStatementsCache = renderDuckRemoteMutationStatements(
-		b.statements,
-		duckRemoteInsertCoalesceMaxRows,
-		duckRemoteInsertCoalesceMaxBytes,
-	)
-	b.renderedStatementBytes = 0
-	for _, stmt := range b.renderedStatementsCache {
-		b.renderedStatementBytes += len(stmt.sql)
-	}
-	b.renderedValid = true
-	return b.renderedStatementsCache
-}
-
-func renderDuckRemoteMutationStatements(
-	statements []string, maxRows int, maxBytes int,
-) []duckRemoteRenderedStatement {
-	if len(statements) == 0 {
-		return nil
-	}
-	if maxRows <= 0 {
-		maxRows = duckRemoteInsertCoalesceMaxRows
-	}
-	if maxBytes <= 0 {
-		maxBytes = duckRemoteInsertCoalesceMaxBytes
-	}
-	out := make([]duckRemoteRenderedStatement, 0, len(statements))
-	var pending duckRemoteInsertGroup
-	hasPending := false
-	flush := func() {
-		if !hasPending {
-			return
-		}
-		out = append(out, pending.RenderedStatement())
-		pending = duckRemoteInsertGroup{}
-		hasPending = false
-	}
-	for i, stmt := range statements {
-		prefix, tuple, ok := splitDuckRemoteSimpleInsert(stmt)
-		if !ok {
-			flush()
-			out = append(out, duckRemoteRenderedStatement{
-				sql:   stmt,
-				start: i,
-				end:   i + 1,
-			})
-			continue
-		}
-		if !hasPending || pending.prefix != prefix {
-			flush()
-			pending = duckRemoteInsertGroup{prefix: prefix, start: i}
-			hasPending = true
-		}
-		if pending.rows > 0 && (pending.rows+1 > maxRows ||
-			pending.bytes+len(", ")+len(tuple) > maxBytes) {
-			flush()
-			pending = duckRemoteInsertGroup{prefix: prefix, start: i}
-			hasPending = true
-		}
-		pending.Append(tuple, i+1)
-	}
-	flush()
-	return out
-}
-
-type duckRemoteInsertGroup struct {
-	prefix string
-	tuples []string
-	rows   int
-	bytes  int
-	start  int
-	end    int
-}
-
-func (g *duckRemoteInsertGroup) Append(tuple string, end int) {
-	if g.rows == 0 {
-		g.bytes = len(g.prefix)
-	} else {
-		g.bytes += len(", ")
-	}
-	g.tuples = append(g.tuples, tuple)
-	g.rows++
-	g.bytes += len(tuple)
-	g.end = end
-}
-
-func (g duckRemoteInsertGroup) RenderedStatement() duckRemoteRenderedStatement {
-	return duckRemoteRenderedStatement{
-		sql:   g.prefix + strings.Join(g.tuples, ", "),
-		start: g.start,
-		end:   g.end,
-	}
-}
-
-func splitDuckRemoteSimpleInsert(stmt string) (string, string, bool) {
-	trimmed := strings.TrimSpace(stmt)
-	if !duckASCIIHasPrefixFold(trimmed, "INSERT INTO ") {
-		return "", "", false
-	}
-	valuesIndex := duckSQLValuesKeywordIndex(trimmed)
-	if valuesIndex < 0 {
-		return "", "", false
-	}
-	prefix := strings.TrimSpace(trimmed[:valuesIndex]) + " VALUES "
-	rest := strings.TrimSpace(trimmed[valuesIndex+len(" VALUES "):])
-	tuple, suffix, ok := duckSQLLeadingParenthesizedExpression(rest)
-	if !ok || strings.TrimSpace(suffix) != "" {
-		return "", "", false
-	}
-	return prefix, tuple, true
-}
-
-func duckSQLValuesKeywordIndex(sqlText string) int {
-	const valuesKeyword = " VALUES "
-	for i := len("INSERT INTO "); i+len(valuesKeyword) <= len(sqlText); i++ {
-		if duckASCIIEqualFoldAt(sqlText, i, valuesKeyword) {
-			return i
-		}
-	}
-	return -1
-}
-
-func duckASCIIHasPrefixFold(s, prefix string) bool {
-	if len(s) < len(prefix) {
-		return false
-	}
-	return duckASCIIEqualFoldAt(s, 0, prefix)
-}
-
-func duckASCIIEqualFoldAt(s string, start int, pattern string) bool {
-	if start < 0 || start+len(pattern) > len(s) {
-		return false
-	}
-	for i := range len(pattern) {
-		if duckASCIIFold(s[start+i]) != duckASCIIFold(pattern[i]) {
-			return false
-		}
-	}
-	return true
-}
-
-func duckASCIIFold(b byte) byte {
-	if b >= 'a' && b <= 'z' {
-		return b - 'a' + 'A'
-	}
-	return b
-}
-
-func duckSQLLeadingParenthesizedExpression(sqlText string) (
-	string, string, bool,
-) {
-	if !strings.HasPrefix(sqlText, "(") {
-		return "", "", false
-	}
-	depth := 0
-	for i := 0; i < len(sqlText); {
-		switch sqlText[i] {
-		case '$':
-			delimiter, ok := duckDollarQuoteDelimiterAt(sqlText, i)
-			if ok {
-				next := strings.Index(sqlText[i+len(delimiter):], delimiter)
-				if next < 0 {
-					return "", "", false
-				}
-				i += len(delimiter) + next + len(delimiter)
-				continue
-			}
-		case '\'':
-			next, ok := duckSingleQuotedLiteralEnd(sqlText, i)
-			if !ok {
-				return "", "", false
-			}
-			i = next
-			continue
-		case '(':
-			depth++
-		case ')':
-			depth--
-			if depth == 0 {
-				return sqlText[:i+1], sqlText[i+1:], true
-			}
-			if depth < 0 {
-				return "", "", false
-			}
-		}
-		i++
-	}
-	return "", "", false
-}
-
-func duckDollarQuoteDelimiterAt(sqlText string, start int) (string, bool) {
-	if start >= len(sqlText) || sqlText[start] != '$' {
-		return "", false
-	}
-	i := start + 1
-	for i < len(sqlText) && duckDollarQuoteTagByte(sqlText[i]) {
-		i++
-	}
-	if i >= len(sqlText) || sqlText[i] != '$' {
-		return "", false
-	}
-	return sqlText[start : i+1], true
-}
-
-func duckDollarQuoteTagByte(b byte) bool {
-	return b == '_' ||
-		('0' <= b && b <= '9') ||
-		('A' <= b && b <= 'Z') ||
-		('a' <= b && b <= 'z')
-}
-
-func duckSingleQuotedLiteralEnd(sqlText string, start int) (int, bool) {
-	for i := start + 1; i < len(sqlText); i++ {
-		if sqlText[i] != '\'' {
-			continue
-		}
-		if i+1 < len(sqlText) && sqlText[i+1] == '\'' {
-			i++
-			continue
-		}
-		return i + 1, true
-	}
-	return 0, false
-}
-
-func isDuckRemoteMutationTimeoutError(err error) bool {
-	if err == nil {
-		return false
-	}
-	return strings.Contains(strings.ToLower(err.Error()), "timeout was reached")
-}
-
-func (s *Sync) execRemoteSQLRetry(ctx context.Context, sqlText string) error {
-	if s.quack != nil {
-		return s.quack.execRemote(ctx, sqlText, true)
-	}
-	return s.execRemoteSQLNoRetry(ctx, sqlText)
-}
-
-func (s *Sync) execRemoteSQLNoRetry(ctx context.Context, sqlText string) error {
-	if s.quack != nil {
-		return s.quack.execRemote(ctx, sqlText, false)
-	}
-	_, err := s.duck.ExecContext(ctx, "FROM "+quackAttachmentName+".query(?)", sqlText)
-	return err
-}
-
-func duckSQLWithArgs(stmt string, args ...any) (string, error) {
-	var b strings.Builder
-	argIndex := 0
-	for _, r := range stmt {
-		if r != '?' {
-			b.WriteRune(r)
-			continue
-		}
-		if argIndex >= len(args) {
-			return "", fmt.Errorf("duckdb remote statement missing argument")
-		}
-		lit, err := duckValueLiteral(args[argIndex])
-		if err != nil {
-			return "", err
-		}
-		b.WriteString(lit)
-		argIndex++
-	}
-	if argIndex != len(args) {
-		return "", fmt.Errorf("duckdb remote statement has unused argument")
-	}
-	return b.String(), nil
-}
-
-func duckValueLiteral(v any) (string, error) {
-	switch value := v.(type) {
-	case nil:
-		return "NULL", nil
-	case string:
-		return duckRemoteStringLiteral(value)
-	case *string:
-		if value == nil {
-			return "NULL", nil
-		}
-		return duckRemoteStringLiteral(*value)
-	case int, int8, int16, int32, int64,
-		uint, uint8, uint16, uint32, uint64,
-		float32, float64:
-		return fmt.Sprint(value), nil
-	case *int:
-		if value == nil {
-			return "NULL", nil
-		}
-		return fmt.Sprint(*value), nil
-	case *int64:
-		if value == nil {
-			return "NULL", nil
-		}
-		return fmt.Sprint(*value), nil
-	case *float64:
-		if value == nil {
-			return "NULL", nil
-		}
-		return fmt.Sprint(*value), nil
-	case bool:
-		if value {
-			return "TRUE", nil
-		}
-		return "FALSE", nil
-	case time.Time:
-		return "TIMESTAMP " + duckLiteral(
-			value.UTC().Format("2006-01-02 15:04:05.999999"),
-		), nil
-	default:
-		// Mirror database/sql's default parameter converter so named kinds
-		// (export.WorktreeRelationship and friends) render the same way on
-		// the remote transport as they bind on the local mirror path.
-		rv := reflect.ValueOf(v)
-		switch rv.Kind() {
-		case reflect.String:
-			return duckRemoteStringLiteral(rv.String())
-		case reflect.Bool:
-			if rv.Bool() {
-				return "TRUE", nil
-			}
-			return "FALSE", nil
-		case reflect.Int, reflect.Int8, reflect.Int16,
-			reflect.Int32, reflect.Int64:
-			return strconv.FormatInt(rv.Int(), 10), nil
-		case reflect.Uint, reflect.Uint8, reflect.Uint16,
-			reflect.Uint32, reflect.Uint64:
-			return strconv.FormatUint(rv.Uint(), 10), nil
-		case reflect.Float32, reflect.Float64:
-			return strconv.FormatFloat(rv.Float(), 'g', -1, 64), nil
-		default:
-			return "", fmt.Errorf("unsupported duckdb remote argument type %T", v)
-		}
-	}
-}
-
-func duckRemoteStringLiteral(s string) (string, error) {
-	s = strings.ReplaceAll(s, "\x00", "")
-	for {
-		var tagBytes [16]byte
-		if _, err := rand.Read(tagBytes[:]); err != nil {
-			return "", fmt.Errorf("generating duckdb string literal tag: %w", err)
-		}
-		tag := "agentsview_" + hex.EncodeToString(tagBytes[:])
-		delimiter := "$" + tag + "$"
-		if strings.Contains(s, delimiter) {
-			continue
-		}
-		return delimiter + s + delimiter, nil
-	}
 }
 
 func (s *Sync) upsertSession(
