+++ internal/duckdb/sync_test.go
@@ -262,94 +262,6 @@ func TestPushSessionBatchLogsAbandonedSessionsAfterContextCancel(
 	assert.Contains(t, gotLogs, "abandoning 2 sessions")
 }
 
-func TestPushSessionBatchBacksOffAndRetriesBatchAfterTimeout(t *testing.T) {
-	ctx := context.Background()
-	sessions := []db.Session{
-		syncSession(
-			"duck-timeout-batch-1", "alpha", "timeout one",
-			"2026-01-10T00:00:00.000Z", 1,
-		),
-		syncSession(
-			"duck-timeout-batch-2", "alpha", "timeout two",
-			"2026-01-10T00:01:00.000Z", 2,
-		),
-	}
-	var result PushResult
-	var pushed []db.Session
-	timeoutErr := fmt.Errorf(
-		"IO Error: Failed to send message: IO Error: Timeout was reached error for HTTP POST to '<url>'",
-	)
-	tryCalls := 0
-	waitCalls := 0
-	pushSingleCalls := 0
-
-	err := pushSessionBatchWith(
-		ctx, sessions, 0, len(sessions), &result, &pushed, nil,
-		func(_ context.Context, got []db.Session) ([]int, error) {
-			tryCalls++
-			assert.Equal(t, sessions, got)
-			if tryCalls == 1 {
-				return nil, timeoutErr
-			}
-			return []int{1, 2}, nil
-		},
-		func(context.Context, db.Session) (int, error) {
-			pushSingleCalls++
-			return 0, fmt.Errorf("individual retry should not run")
-		},
-		func(context.Context) error {
-			waitCalls++
-			return nil
-		},
-	)
-	require.NoError(t, err)
-
-	assert.Equal(t, 2, tryCalls)
-	assert.Equal(t, 1, waitCalls)
-	assert.Equal(t, 0, pushSingleCalls)
-	assert.Equal(t, 2, result.SessionsPushed)
-	assert.Equal(t, 3, result.MessagesPushed)
-	assert.Equal(t, 0, result.Errors)
-	assert.Equal(t, sessions, pushed)
-}
-
-func TestPushSessionBatchReturnsRepeatedTimeoutWithoutIndividualRetry(t *testing.T) {
-	ctx := context.Background()
-	sessions := []db.Session{
-		syncSession(
-			"duck-timeout-batch-repeat", "alpha", "timeout",
-			"2026-01-10T00:00:00.000Z", 1,
-		),
-	}
-	var result PushResult
-	var pushed []db.Session
-	timeoutErr := fmt.Errorf(
-		"IO Error: Failed to send message: IO Error: Timeout was reached error for HTTP POST to '<url>'",
-	)
-	tryCalls := 0
-	pushSingleCalls := 0
-
-	err := pushSessionBatchWith(
-		ctx, sessions, 0, len(sessions), &result, &pushed, nil,
-		func(context.Context, []db.Session) ([]int, error) {
-			tryCalls++
-			return nil, timeoutErr
-		},
-		func(context.Context, db.Session) (int, error) {
-			pushSingleCalls++
-			return 0, fmt.Errorf("individual retry should not run")
-		},
-		func(context.Context) error { return nil },
-	)
-	require.Error(t, err)
-
-	assert.True(t, isDuckRemoteMutationTimeoutError(err))
-	assert.Equal(t, 2, tryCalls)
-	assert.Equal(t, 0, pushSingleCalls)
-	assert.Equal(t, 0, result.Errors)
-	assert.Empty(t, pushed)
-}
-
 func TestSyncPushReportsSessionDiagnosticsByAgent(t *testing.T) {
 	ctx := context.Background()
 	local := newLocalDB(t)
@@ -430,793 +342,6 @@ func TestSyncPushReportsProgressAcrossBatchBoundaries(t *testing.T) {
 	assert.Equal(t, count, progress[duckSessionPushBatchSize].SessionsDone)
 	assert.Equal(t, count, progress[duckSessionPushBatchSize].SessionsTotal)
 	assert.Equal(t, count, progress[duckSessionPushBatchSize].MessagesDone)
-}
-
-func TestDuckValueLiteralFormatsTimestampWithoutZone(t *testing.T) {
-	got, err := duckValueLiteral(time.Date(
-		2026, time.January, 10, 3, 4, 5, 123456789, time.UTC,
-	))
-	require.NoError(t, err)
-
-	assert.Equal(t, "TIMESTAMP '2026-01-10 03:04:05.123456'", got)
-}
-
-func TestDuckSQLWithArgsExecutesQuotedMultilineString(t *testing.T) {
-	ctx := context.Background()
-	duck := openTestDuckDB(t)
-	want := "first line\nquoted ' value\ncontains $$ delimiter text"
-
-	stmt, err := duckSQLWithArgs(`SELECT ?`, want)
-	require.NoError(t, err)
-
-	var got string
-	require.NoError(t, duck.QueryRowContext(ctx, stmt).Scan(&got))
-	assert.Equal(t, want, got)
-}
-
-func TestDuckSQLWithArgsStripsNULFromStringLiteral(t *testing.T) {
-	ctx := context.Background()
-	duck := openTestDuckDB(t)
-
-	stmt, err := duckSQLWithArgs(`SELECT ?`, "before\x00after")
-	require.NoError(t, err)
-
-	var got string
-	require.NoError(t, duck.QueryRowContext(ctx, stmt).Scan(&got))
-	assert.Equal(t, "beforeafter", got)
-}
-
-func TestDuckSQLWithArgsExecutesStringPointer(t *testing.T) {
-	ctx := context.Background()
-	duck := openTestDuckDB(t)
-	want := "pinned note\nquoted ' value"
-
-	stmt, err := duckSQLWithArgs(`SELECT ?`, &want)
-	require.NoError(t, err)
-
-	var got string
-	require.NoError(t, duck.QueryRowContext(ctx, stmt).Scan(&got))
-	assert.Equal(t, want, got)
-}
-
-func TestExecDuckRemoteMutationBatchCoalescesSuccessfulBatch(t *testing.T) {
-	ctx := context.Background()
-	batch := &duckRemoteMutationBatch{}
-	_, err := batch.ExecContext(ctx, `INSERT INTO remote_test VALUES (?)`, 1)
-	require.NoError(t, err)
-	_, err = batch.ExecContext(ctx, `DELETE FROM remote_test WHERE id = ?`, 2)
-	require.NoError(t, err)
-	var calls []string
-
-	err = execDuckRemoteMutationBatch(
-		ctx,
-		func(_ context.Context, sqlText string) error {
-			calls = append(calls, sqlText)
-			return nil
-		},
-		"test batch",
-		batch,
-		true,
-	)
-	require.NoError(t, err)
-
-	require.Len(t, calls, 1)
-	assert.Equal(t, `BEGIN TRANSACTION;
-INSERT INTO remote_test VALUES (1);
-DELETE FROM remote_test WHERE id = 2;
-COMMIT`, calls[0])
-}
-
-func TestExecDuckRemoteMutationBatchCoalescedFailureRollsBack(t *testing.T) {
-	ctx := context.Background()
-	batch := &duckRemoteMutationBatch{}
-	_, err := batch.ExecContext(ctx, `INSERT INTO remote_test VALUES (?)`, 1)
-	require.NoError(t, err)
-	var calls []string
-
-	err = execDuckRemoteMutationBatch(
-		ctx,
-		func(_ context.Context, sqlText string) error {
-			calls = append(calls, sqlText)
-			if sqlText == "ROLLBACK" {
-				return nil
-			}
-			return fmt.Errorf("batch failed")
-		},
-		"test batch",
-		batch,
-		true,
-	)
-
-	require.ErrorContains(t, err, "batch failed")
-	require.Len(t, calls, 2)
-	assert.Contains(t, calls[0], "BEGIN TRANSACTION")
-	assert.Equal(t, "ROLLBACK", calls[1])
-}
-
-func TestExecDuckRemoteMutationBatchCoalescedTimeoutSkipsRollback(
-	t *testing.T,
-) {
-	ctx := context.Background()
-	batch := &duckRemoteMutationBatch{}
-	_, err := batch.ExecContext(ctx, `INSERT INTO remote_test VALUES (?)`, 1)
-	require.NoError(t, err)
-	var calls []string
-
-	err = execDuckRemoteMutationBatch(
-		ctx,
-		func(_ context.Context, sqlText string) error {
-			calls = append(calls, sqlText)
-			return fmt.Errorf(
-				"IO Error: Failed to send message: IO Error: Timeout was reached error for HTTP POST to '<url>'",
-			)
-		},
-		"test batch",
-		batch,
-		true,
-	)
-
-	require.Error(t, err)
-	assert.True(t, isDuckRemoteMutationTimeoutError(err))
-	assert.Equal(t, []string{batch.transactionSQL()}, calls)
-}
-
-func TestExecDuckRemoteMutationBatchCoalescedRollbackFailureIsWrapped(
-	t *testing.T,
-) {
-	ctx := context.Background()
-	batch := &duckRemoteMutationBatch{}
-	_, err := batch.ExecContext(ctx, `INSERT INTO remote_test VALUES (?)`, 1)
-	require.NoError(t, err)
-
-	err = execDuckRemoteMutationBatch(
-		ctx,
-		func(_ context.Context, sqlText string) error {
-			if sqlText == "ROLLBACK" {
-				return fmt.Errorf("rollback failed")
-			}
-			return fmt.Errorf("batch failed")
-		},
-		"test batch",
-		batch,
-		true,
-	)
-
-	require.ErrorContains(t, err, "batch failed")
-	assert.ErrorContains(t, err, "rollback test batch")
-	assert.ErrorContains(t, err, "rollback failed")
-}
-
-func TestDuckRemoteMutationDefaultBudgetFitsQuackTimeout(t *testing.T) {
-	assert.GreaterOrEqual(t, duckRemoteMutationCoalesceMaxBytes, 2<<20)
-	assert.LessOrEqual(t, duckRemoteMutationCoalesceMaxBytes, 4<<20)
-}
-
-func TestDuckRemoteMutationBatchByteAccountingMatchesRenderedTransaction(
-	t *testing.T,
-) {
-	ctx := context.Background()
-	batch := &duckRemoteMutationBatch{}
-	_, err := batch.ExecContext(ctx,
-		`INSERT INTO remote_test VALUES (?, ?)`, 1, "first",
-	)
-	require.NoError(t, err)
-	_, err = batch.ExecContext(ctx,
-		`INSERT INTO remote_test VALUES (?, ?)`, 2, "second",
-	)
-	require.NoError(t, err)
-
-	assert.Equal(t, len(batch.transactionSQL()), batch.transactionBytes())
-}
-
-func TestDuckRemoteMutationBatchCoalescesAdjacentInsertStatements(
-	t *testing.T,
-) {
-	ctx := context.Background()
-	batch := &duckRemoteMutationBatch{}
-	_, err := batch.ExecContext(ctx,
-		`DELETE FROM remote_test WHERE session_id = ?`, "session-1",
-	)
-	require.NoError(t, err)
-	for i := range 3 {
-		_, err = batch.ExecContext(ctx,
-			`INSERT INTO remote_test (session_id, ordinal) VALUES (?, ?)`,
-			"session-1", i,
-		)
-		require.NoError(t, err)
-	}
-
-	sqlText := batch.transactionSQL()
-
-	assert.Equal(t, 1, strings.Count(sqlText, "INSERT INTO remote_test"))
-	assert.Contains(t, sqlText,
-		"INSERT INTO remote_test (session_id, ordinal) VALUES ("+
-			"$agentsview_")
-	assert.Contains(t, sqlText, ", 0), (")
-	assert.Contains(t, sqlText, ", 1), (")
-	assert.Contains(t, sqlText, ", 2);")
-	assert.Less(t, batch.transactionBytes(), len(strings.Join(batch.statements, ";\n")))
-}
-
-func TestDuckRemoteMutationBatchInsertCoalescingSplitsByRowLimit(
-	t *testing.T,
-) {
-	ctx := context.Background()
-	batch := &duckRemoteMutationBatch{}
-	for i := range 257 {
-		_, err := batch.ExecContext(ctx,
-			`INSERT INTO remote_test (session_id, ordinal) VALUES (?, ?)`,
-			"session-1", i,
-		)
-		require.NoError(t, err)
-	}
-
-	sqlText := batch.transactionSQL()
-
-	assert.Equal(t, 2, strings.Count(sqlText, "INSERT INTO remote_test"))
-}
-
-func TestDuckRemoteMutationBatchCoalescedInsertMatchesPerRowExecution(
-	t *testing.T,
-) {
-	ctx := context.Background()
-	coalescedDB := openTestDuckDB(t)
-	perRowDB := openTestDuckDB(t)
-	for _, conn := range []*sql.DB{coalescedDB, perRowDB} {
-		_, err := conn.ExecContext(ctx,
-			`CREATE TABLE remote_test (
-				session_id VARCHAR,
-				ordinal INTEGER,
-				content VARCHAR
-			)`,
-		)
-		require.NoError(t, err)
-		_, err = conn.ExecContext(ctx,
-			`INSERT INTO remote_test VALUES ('session-1', 99, 'stale')`,
-		)
-		require.NoError(t, err)
-	}
-	batch := &duckRemoteMutationBatch{}
-	_, err := batch.ExecContext(ctx,
-		`DELETE FROM remote_test WHERE session_id = ?`, "session-1",
-	)
-	require.NoError(t, err)
-	for _, row := range []struct {
-		ordinal int
-		content string
-	}{
-		{ordinal: 0, content: "first"},
-		{ordinal: 1, content: "second"},
-		{ordinal: 2, content: "third"},
-	} {
-		_, err = batch.ExecContext(ctx,
-			`INSERT INTO remote_test (session_id, ordinal, content) VALUES (?, ?, ?)`,
-			"session-1", row.ordinal, row.content,
-		)
-		require.NoError(t, err)
-	}
-
-	_, err = coalescedDB.ExecContext(ctx, batch.transactionSQL())
-	require.NoError(t, err)
-	require.NoError(t, execDuckRemoteMutationBatch(
-		ctx,
-		func(ctx context.Context, sqlText string) error {
-			_, err := perRowDB.ExecContext(ctx, sqlText)
-			return err
-		},
-		"per-row equivalence",
-		batch,
-		false,
-	))
-
-	assert.Equal(
-		t,
-		readRemoteTestRows(t, ctx, perRowDB),
-		readRemoteTestRows(t, ctx, coalescedDB),
-	)
-}
-
-type remoteTestRow struct {
-	SessionID string
-	Ordinal   int
-	Content   string
-}
-
-func readRemoteTestRows(
-	t *testing.T, ctx context.Context, conn *sql.DB,
-) []remoteTestRow {
-	t.Helper()
-	rows, err := conn.QueryContext(ctx,
-		`SELECT session_id, ordinal, content
-		 FROM remote_test
-		 ORDER BY session_id, ordinal, content`,
-	)
-	require.NoError(t, err)
-	defer rows.Close()
-	var out []remoteTestRow
-	for rows.Next() {
-		var row remoteTestRow
-		require.NoError(t, rows.Scan(
-			&row.SessionID, &row.Ordinal, &row.Content,
-		))
-		out = append(out, row)
-	}
-	require.NoError(t, rows.Err())
-	return out
-}
-
-func TestDuckRemoteMutationBatchFallbackKeepsPerRowInsertAttribution(
-	t *testing.T,
-) {
-	ctx := context.Background()
-	batch := &duckRemoteMutationBatch{}
-	_, err := batch.ExecContext(ctx, `INSERT INTO remote_test VALUES (?)`, 1)
-	require.NoError(t, err)
-	_, err = batch.ExecContext(ctx, `INSERT INTO remote_test VALUES (?)`, 2)
-	require.NoError(t, err)
-	var coalescedCalls []string
-	var statementCalls []string
-
-	err = execDuckRemoteMutationBatchWithStatementFallback(
-		ctx,
-		func(_ context.Context, sqlText string) error {
-			coalescedCalls = append(coalescedCalls, sqlText)
-			if sqlText == "ROLLBACK" {
-				return nil
-			}
-			return fmt.Errorf("coalesced failure")
-		},
-		func(_ context.Context, sqlText string) error {
-			statementCalls = append(statementCalls, sqlText)
-			if sqlText == "INSERT INTO remote_test VALUES (2)" {
-				return fmt.Errorf("poison row")
-			}
-			return nil
-		},
-		"test session",
-		batch,
-	)
-	require.Error(t, err)
-
-	require.NotEmpty(t, coalescedCalls)
-	assert.Equal(t, 1, strings.Count(coalescedCalls[0], "INSERT INTO remote_test"))
-	assert.Contains(t, err.Error(), "execute test session statement 2/2")
-	assert.Equal(t, []string{
-		"BEGIN TRANSACTION",
-		"INSERT INTO remote_test VALUES (1)",
-		"INSERT INTO remote_test VALUES (2)",
-		"ROLLBACK",
-	}, statementCalls)
-}
-
-func TestDuckRemoteMutationBatchCombinedTransactionBytesUsesCachedRendering(
-	t *testing.T,
-) {
-	ctx := context.Background()
-	batch := &duckRemoteMutationBatch{}
-	for i := range 50 {
-		_, err := batch.ExecContext(ctx,
-			`INSERT INTO remote_test VALUES (?, ?)`, i, strings.Repeat("x", 32),
-		)
-		require.NoError(t, err)
-	}
-	next := &duckRemoteMutationBatch{}
-	_, err := next.ExecContext(ctx,
-		`INSERT INTO remote_test VALUES (?, ?)`, 51, strings.Repeat("y", 32),
-	)
-	require.NoError(t, err)
-
-	require.Positive(t, batch.transactionBytes())
-	require.Positive(t, next.transactionBytes())
-	allocs := testing.AllocsPerRun(100, func() {
-		_ = batch.combinedTransactionBytes(next)
-	})
-
-	assert.Zero(t, allocs)
-}
-
-func TestDuckRemoteMutationBatchAppendBatchPreservesRenderedCache(
-	t *testing.T,
-) {
-	ctx := context.Background()
-	current := &duckRemoteMutationBatch{}
-	first := &duckRemoteMutationBatch{}
-	_, err := first.ExecContext(ctx,
-		`INSERT INTO remote_test VALUES (?, ?)`, 1, strings.Repeat("a", 32),
-	)
-	require.NoError(t, err)
-	second := &duckRemoteMutationBatch{}
-	_, err = second.ExecContext(ctx,
-		`INSERT INTO remote_test VALUES (?, ?)`, 2, strings.Repeat("b", 32),
-	)
-	require.NoError(t, err)
-	third := &duckRemoteMutationBatch{}
-	_, err = third.ExecContext(ctx,
-		`INSERT INTO remote_test VALUES (?, ?)`, 3, strings.Repeat("c", 32),
-	)
-	require.NoError(t, err)
-
-	current.appendBatch(first)
-	require.Positive(t, current.combinedTransactionBytes(second))
-	current.appendBatch(second)
-	require.Positive(t, third.transactionBytes())
-	allocs := testing.AllocsPerRun(100, func() {
-		_ = current.combinedTransactionBytes(third)
-	})
-
-	assert.Zero(t, allocs)
-}
-
-func TestDuckRemoteMutationBatchAppendPreservesRenderedCache(t *testing.T) {
-	ctx := context.Background()
-	current := &duckRemoteMutationBatch{}
-	_, err := current.ExecContext(ctx,
-		`INSERT INTO remote_test VALUES (?, ?)`, 1, strings.Repeat("x", 32),
-	)
-	require.NoError(t, err)
-	next := &duckRemoteMutationBatch{}
-	_, err = next.ExecContext(ctx,
-		`INSERT INTO remote_test VALUES (?, ?)`, 2, strings.Repeat("y", 32),
-	)
-	require.NoError(t, err)
-
-	require.Positive(t, current.combinedTransactionBytes(next))
-	require.True(t, current.renderedValid)
-	current.appendBatch(next)
-
-	assert.True(t, current.renderedValid)
-	assert.Len(t, current.rendered(), 2)
-}
-
-func TestSplitDuckRemoteSimpleInsertAllocatesOnlyReturnedPrefix(
-	t *testing.T,
-) {
-	literal, err := duckRemoteStringLiteral(strings.Repeat("x", 4096))
-	require.NoError(t, err)
-	stmt := "INSERT INTO remote_test VALUES (" + literal + ")"
-	prefix, tuple, ok := splitDuckRemoteSimpleInsert(stmt)
-	require.True(t, ok)
-	assert.Equal(t, "INSERT INTO remote_test VALUES ", prefix)
-	assert.Contains(t, tuple, strings.Repeat("x", 128))
-
-	allocs := testing.AllocsPerRun(100, func() {
-		_, _, _ = splitDuckRemoteSimpleInsert(stmt)
-	})
-
-	assert.LessOrEqual(t, allocs, 1.0)
-}
-
-func TestAppendDuckRemoteMutationBatchFlushesBeforeByteBudget(t *testing.T) {
-	ctx := context.Background()
-	first := &duckRemoteMutationBatch{}
-	_, err := first.ExecContext(ctx,
-		`INSERT INTO remote_test VALUES (?, ?)`, 1, "first",
-	)
-	require.NoError(t, err)
-	second := &duckRemoteMutationBatch{}
-	_, err = second.ExecContext(ctx,
-		`INSERT INTO remote_test VALUES (?, ?)`, 2, "second",
-	)
-	require.NoError(t, err)
-	maxBytes := first.combinedTransactionBytes(second) - 1
-	current := &duckRemoteMutationBatch{}
-	var calls []string
-	exec := func(_ context.Context, sqlText string) error {
-		calls = append(calls, sqlText)
-		return nil
-	}
-
-	current, err = appendDuckRemoteMutationBatch(
-		ctx, exec, "test batch", current, first, maxBytes,
-	)
-	require.NoError(t, err)
-	current, err = appendDuckRemoteMutationBatch(
-		ctx, exec, "test batch", current, second, maxBytes,
-	)
-	require.NoError(t, err)
-	require.Len(t, calls, 1)
-	assert.Contains(t, calls[0], "$")
-	assert.Contains(t, calls[0], "first")
-	assert.NotContains(t, calls[0], "second")
-
-	require.NoError(t, execDuckRemoteMutationBatch(
-		ctx, exec, "test batch", current, true,
-	))
-	require.Len(t, calls, 2)
-	assert.Contains(t, calls[1], "second")
-}
-
-func TestExecDuckRemoteMutationBatchOversizeUsesStatementModeTransaction(
-	t *testing.T,
-) {
-	ctx := context.Background()
-	oversize := &duckRemoteMutationBatch{}
-	_, err := oversize.ExecContext(ctx,
-		`INSERT INTO remote_test VALUES (?, ?)`, 1, "aaaaa",
-	)
-	require.NoError(t, err)
-	_, err = oversize.ExecContext(ctx,
-		`INSERT INTO remote_test VALUES (?, ?)`, 2, "bbbbb",
-	)
-	require.NoError(t, err)
-	_, err = oversize.ExecContext(ctx,
-		`INSERT INTO remote_test VALUES (?, ?)`, 3, "ccccc",
-	)
-	require.NoError(t, err)
-	oneStatement := &duckRemoteMutationBatch{statements: oversize.statements[:1]}
-	maxBytes := oneStatement.transactionBytes()
-	var coalescedCalls []string
-	var statementCalls []string
-
-	err = execDuckRemoteMutationBatchOversizeWithStatementFallback(
-		ctx,
-		func(_ context.Context, sqlText string) error {
-			coalescedCalls = append(coalescedCalls, sqlText)
-			return nil
-		},
-		func(_ context.Context, sqlText string) error {
-			statementCalls = append(statementCalls, sqlText)
-			return nil
-		},
-		"oversize session",
-		oversize,
-		maxBytes,
-	)
-	require.NoError(t, err)
-
-	assert.Empty(t, coalescedCalls)
-	require.Len(t, statementCalls, 3)
-	assert.Equal(t, "BEGIN TRANSACTION", statementCalls[0])
-	assert.Contains(t, statementCalls[1], "INSERT INTO remote_test VALUES (1, $")
-	assert.Contains(t, statementCalls[1], "), (2, $")
-	assert.Contains(t, statementCalls[1], "), (3, $")
-	assert.Equal(t, "COMMIT", statementCalls[2])
-}
-
-func TestAppendDuckRemoteMutationBatchRejectsOversizeSessionAfterFlush(
-	t *testing.T,
-) {
-	ctx := context.Background()
-	current := &duckRemoteMutationBatch{}
-	_, err := current.ExecContext(ctx,
-		`INSERT INTO remote_test VALUES (?, ?)`, 1, "current",
-	)
-	require.NoError(t, err)
-	oversize := &duckRemoteMutationBatch{}
-	_, err = oversize.ExecContext(ctx,
-		`INSERT INTO remote_test VALUES (?, ?)`, 2, "bbbbb",
-	)
-	require.NoError(t, err)
-	_, err = oversize.ExecContext(ctx,
-		`INSERT INTO remote_test VALUES (?, ?)`, 3, "ccccc",
-	)
-	require.NoError(t, err)
-	oneStatement := &duckRemoteMutationBatch{statements: oversize.statements[:1]}
-	maxBytes := oneStatement.transactionBytes()
-	var calls []string
-
-	next, err := appendDuckRemoteMutationBatch(
-		ctx,
-		func(_ context.Context, sqlText string) error {
-			calls = append(calls, sqlText)
-			return nil
-		},
-		"oversize session",
-		current,
-		oversize,
-		maxBytes,
-	)
-	require.ErrorContains(t, err, "exceeds remote mutation coalesce budget")
-
-	assert.Equal(t, 0, next.Len())
-	require.Len(t, calls, 1)
-	assert.Equal(t, current.transactionSQL(), calls[0])
-}
-
-func TestExecDuckRemoteMutationBatchStatementModePinpointsFailure(t *testing.T) {
-	ctx := context.Background()
-	batch := &duckRemoteMutationBatch{}
-	_, err := batch.ExecContext(ctx, `INSERT INTO remote_test VALUES (?)`, 1)
-	require.NoError(t, err)
-	_, err = batch.ExecContext(ctx, `DELETE FROM remote_test WHERE id = ?`, 2)
-	require.NoError(t, err)
-	var calls []string
-
-	err = execDuckRemoteMutationBatch(
-		ctx,
-		func(_ context.Context, sqlText string) error {
-			calls = append(calls, sqlText)
-			if strings.HasPrefix(sqlText, "DELETE") {
-				return fmt.Errorf("delete failed")
-			}
-			return nil
-		},
-		"test session",
-		batch,
-		false,
-	)
-	require.Error(t, err)
-
-	assert.Contains(t, err.Error(), "execute test session statement 2/2")
-	assert.Equal(t, []string{
-		"BEGIN TRANSACTION",
-		"INSERT INTO remote_test VALUES (1)",
-		"DELETE FROM remote_test WHERE id = 2",
-		"ROLLBACK",
-	}, calls)
-}
-
-func TestExecDuckRemoteMutationBatchFallbackLocalizesNonStaleFailure(t *testing.T) {
-	ctx := context.Background()
-	batch := &duckRemoteMutationBatch{}
-	_, err := batch.ExecContext(ctx, `INSERT INTO remote_test VALUES (?)`, 1)
-	require.NoError(t, err)
-	_, err = batch.ExecContext(ctx, `DELETE FROM remote_test WHERE id = ?`, 2)
-	require.NoError(t, err)
-	var coalescedCalls []string
-	var statementCalls []string
-
-	err = execDuckRemoteMutationBatchWithStatementFallback(
-		ctx,
-		func(_ context.Context, sqlText string) error {
-			coalescedCalls = append(coalescedCalls, sqlText)
-			if sqlText == "ROLLBACK" {
-				return nil
-			}
-			return fmt.Errorf("coalesced failure")
-		},
-		func(_ context.Context, sqlText string) error {
-			statementCalls = append(statementCalls, sqlText)
-			if strings.HasPrefix(sqlText, "DELETE") {
-				return fmt.Errorf("delete failed")
-			}
-			return nil
-		},
-		"test session",
-		batch,
-	)
-	require.Error(t, err)
-
-	assert.Equal(t, []string{batch.transactionSQL(), "ROLLBACK"}, coalescedCalls)
-	assert.Contains(t, err.Error(), "execute test session statement 2/2")
-	assert.Equal(t, []string{
-		"BEGIN TRANSACTION",
-		"INSERT INTO remote_test VALUES (1)",
-		"DELETE FROM remote_test WHERE id = 2",
-		"ROLLBACK",
-	}, statementCalls)
-}
-
-func TestExecDuckRemoteMutationBatchFallbackDoesNotRetryStaleFailure(t *testing.T) {
-	ctx := context.Background()
-	batch := &duckRemoteMutationBatch{}
-	_, err := batch.ExecContext(ctx, `INSERT INTO remote_test VALUES (?)`, 1)
-	require.NoError(t, err)
-	var statementCalls []string
-
-	err = execDuckRemoteMutationBatchWithStatementFallback(
-		ctx,
-		func(_ context.Context, sqlText string) error {
-			if sqlText == "ROLLBACK" {
-				return nil
-			}
-			return fmt.Errorf("Invalid Input Error: Invalid connection id")
-		},
-		func(_ context.Context, sqlText string) error {
-			statementCalls = append(statementCalls, sqlText)
-			return nil
-		},
-		"test session",
-		batch,
-	)
-	require.ErrorContains(t, err, "Invalid connection id")
-	assert.Empty(t, statementCalls)
-}
-
-func TestExecDuckRemoteMutationBatchFallbackDoesNotRetryTimeout(
-	t *testing.T,
-) {
-	ctx := context.Background()
-	batch := &duckRemoteMutationBatch{}
-	_, err := batch.ExecContext(ctx, `INSERT INTO remote_test VALUES (?)`, 1)
-	require.NoError(t, err)
-	var coalescedCalls []string
-	var statementCalls []string
-
-	err = execDuckRemoteMutationBatchWithStatementFallback(
-		ctx,
-		func(_ context.Context, sqlText string) error {
-			coalescedCalls = append(coalescedCalls, sqlText)
-			return fmt.Errorf(
-				"IO Error: Failed to send message: IO Error: Timeout was reached error for HTTP POST to '<url>'",
-			)
-		},
-		func(_ context.Context, sqlText string) error {
-			statementCalls = append(statementCalls, sqlText)
-			return nil
-		},
-		"test session",
-		batch,
-	)
-	require.Error(t, err)
-
-	assert.True(t, isDuckRemoteMutationTimeoutError(err))
-	assert.Equal(t, []string{batch.transactionSQL()}, coalescedCalls)
-	assert.Empty(t, statementCalls)
-}
-
-func TestDuckValueLiteralFormatsNullableNumericPointers(t *testing.T) {
-	score := 88
-	fileSize := int64(4096)
-	contextPressure := 0.875
-
-	tests := []struct {
-		name string
-		in   any
-		want string
-	}{
-		{name: "int pointer", in: &score, want: "88"},
-		{name: "int64 pointer", in: &fileSize, want: "4096"},
-		{name: "float64 pointer", in: &contextPressure, want: "0.875"},
-		{name: "nil int pointer", in: (*int)(nil), want: "NULL"},
-		{name: "nil int64 pointer", in: (*int64)(nil), want: "NULL"},
-		{name: "nil float64 pointer", in: (*float64)(nil), want: "NULL"},
-	}
-
-	for _, tt := range tests {
-		t.Run(tt.name, func(t *testing.T) {
-			got, err := duckValueLiteral(tt.in)
-			require.NoError(t, err)
-			assert.Equal(t, tt.want, got)
-		})
-	}
-}
-
-func TestDuckSQLWithArgsExecutesNamedStringKinds(t *testing.T) {
-	ctx := context.Background()
-	duck := openTestDuckDB(t)
-
-	stmt, err := duckSQLWithArgs(
-		`SELECT ?, ?, ?`,
-		export.WorktreeLinked,
-		export.CheckoutBranch,
-		export.ProjectResolutionAmbiguous,
-	)
-	require.NoError(t, err)
-
-	var relationship, checkout, resolution string
-	require.NoError(t, duck.QueryRowContext(ctx, stmt).
-		Scan(&relationship, &checkout, &resolution))
-	assert.Equal(t, string(export.WorktreeLinked), relationship)
-	assert.Equal(t, string(export.CheckoutBranch), checkout)
-	assert.Equal(t, string(export.ProjectResolutionAmbiguous), resolution)
-}
-
-func TestDuckValueLiteralFormatsNamedScalarKinds(t *testing.T) {
-	type namedInt int32
-	type namedUint uint16
-	type namedFloat float64
-	type namedBool bool
-
-	tests := []struct {
-		name string
-		in   any
-		want string
-	}{
-		{name: "named int", in: namedInt(-7), want: "-7"},
-		{name: "named uint", in: namedUint(42), want: "42"},
-		{name: "named float", in: namedFloat(0.875), want: "0.875"},
-		{name: "named bool true", in: namedBool(true), want: "TRUE"},
-		{name: "named bool false", in: namedBool(false), want: "FALSE"},
-	}
-
-	for _, tt := range tests {
-		t.Run(tt.name, func(t *testing.T) {
-			got, err := duckValueLiteral(tt.in)
-			require.NoError(t, err)
-			assert.Equal(t, tt.want, got)
-		})
-	}
 }
 
 func TestSessionFingerprintsStoreDigestOnly(t *testing.T) {
